auth-service/requirements.txt
fastapi
uvicorn
motor
pydantic[email]
pydantic-settings
python-jose[cryptography]
bcrypt==4.0.1
python-multipart
email-validator
